[refactor] 사진 resize form 공통 함수 생성
