+++ src/components/setting/UpdateProfile.tsx
@@ -8,7 +8,6 @@ import {
 import {useNavigation} from '@react-navigation/native';
 import {RootStackNavigationProp} from '~/App';
 import BackView from '~/components/common/BackView';
-import ImageResizer from '@bam.tech/react-native-image-resizer';
 import {TouchableOpacity} from 'react-native';
 import {useUpdateUserInfo} from '~/api/queries/auth';
 import {showToast} from '~/components/common/modal/toastConfig';
@@ -31,6 +30,7 @@ import {
   FONT_NAME,
 } from '../common/style';
 import {GoogleLogoIcon, KakaoLogoIcon, NaverLogoIcon} from '../common/icon';
+import {changeImageSize} from '~/utils/resizeImage';
 
 type InitProfile = {
   favoriteArt: string;
@@ -67,7 +67,6 @@ const UpdateProfile: React.FC<UpdateProfileProps> = ({
   const [imageUri, setImageUri] = useState<string | undefined>(
     initProfile.profile,
   );
-  const [createFormData, setCreateFormData] = useState<FormData | null>(null);
   const [isVerified, setIsVerified] = useState<boolean>(true);
   const [isLoadingOpen, setIsLoadingOpen] = useState<boolean>(false);
   const {
@@ -76,13 +75,7 @@ const UpdateProfile: React.FC<UpdateProfileProps> = ({
     isError,
     isSuccess,
     data: resData,
-  } = useUpdateUserInfo(createFormData);
-
-  useEffect(() => {
-    if (createFormData !== null) {
-      updateUserInfo();
-    }
-  }, [createFormData]);
+  } = useUpdateUserInfo();
 
   useEffect(() => {
     if (isError) {
@@ -108,7 +101,6 @@ const UpdateProfile: React.FC<UpdateProfileProps> = ({
         alarm3: data.alarm3,
         providerType: data.providerType,
       });
-      setCreateFormData(null);
       showToast(messages.successMsg);
       if (navigateTo === 'back') {
         navigation.goBack();
@@ -134,31 +126,11 @@ const UpdateProfile: React.FC<UpdateProfileProps> = ({
     formData.append('nickname', nicknameKeyword);
     formData.append('favoriteArt', art);
 
-    const isImage = imageUri?.search('file://');
-    if (isImage !== undefined && isImage !== -1) {
-      const resizedImage = await ImageResizer.createResizedImage(
-        imageUri ?? '', // path
-        640, // width
-        640, // height
-        'JPEG', // format
-        100, // quality
-        undefined, // rotation
-        // uploadFileName, // outputPath
-        undefined, // keepMeta,
-        undefined, // options => object
-      );
-      const uri = resizedImage.uri;
-      const filename = uri.split('/').pop();
-      const match = /\.(\w+)$/.exec(filename || '');
-      const type = match ? `image/${match[1]}` : `image`;
-
-      formData.append('profile', {
-        name: filename,
-        type,
-        uri: uri,
-      });
+    if (imageUri && imageUri.indexOf('file:///') !== -1) {
+      const resultResizedImage = await changeImageSize(imageUri);
+      formData.append('profile', resultResizedImage);
     }
-    setCreateFormData(formData);
+    updateUserInfo(formData);
   };
 
   return (
